move serialization code, again…
